fix(lottogame, controller, outputview): 총 수익률 변수의 데이터 타입을 실수로 변경

+++ src/main/java/controller/Controller.java
@@ -66,7 +66,7 @@ public class Controller {
     }
 
     private void showLottoGameResult() {
-        long earningsRate = lottoGame.calculateEarningsRate(lottoResult, lottoMachine);
+        double earningsRate = lottoGame.calculateEarningsRate(lottoResult, lottoMachine);
         OutputView.printLottoResult(lottoResult);
         OutputView.printEarningsRate(earningsRate);
     }
